[REFACT] getBasicURL, getReportURl 합치기 #NADA-52
